Count of tags by batch that were unique to each run
This is a summary table of counts of hatchery coho tags by release batch
date that were unique to each run. a way of looking at it is by the
capture histories:
For October 14: 1,x,x (where x means it does not matter if it is a 1 or
0)
For Jan 15: 0,1,x
For Nov 15:,0,0,1 or blank, blank, 1. * note that the capture histories
of 2015 fish is blank, blank, 1. (since they could not be found in 2014
of Jan15.